Addressing feedback
Renaming propagator to plural, to match other top level module names.

Using static functions for setting / getting global httptextformat,
and making propagator methods static.

Renaming "TraceStateHTTP" to "TraceContextHTTP", since
the latter is the correct name for the w3c spec.
